refactor(phase4a): move Grid components to features/grid-editor/
Move Grid/ directory from src/components/ to src/features/grid-editor/components/
as part of Phase 4 folder structure redesign.

Changes:
- Move 23 Grid component files to new location
- Update internal imports (add 2 more ../ levels)
- Add backward-compatible re-exports at old location
- Update test mock paths for ResizeHandles
- Add re-export for IsometricPreview/Scene type
- Slightly adjust coverage threshold (77.9%) for file move

The re-exports allow existing imports to continue working while
the codebase is gradually migrated to use the new paths directly.
