+++ src/components/Grid/index.ts
@@ -1,0 +1,10 @@
+// Re-export from new location for backward compatibility
+// TODO: Update imports to use 'src/features/grid-editor/components/Grid' directly
+export { Grid } from '../../features/grid-editor/components/Grid';
+export { Bin } from '../../features/grid-editor/components/Grid/Bin';
+export { GridCanvas } from '../../features/grid-editor/components/Grid/GridCanvas';
+export { Overlay } from '../../features/grid-editor/components/Grid/Overlay';
+export { ResizeHandle } from '../../features/grid-editor/components/Grid/ResizeHandle';
+export { ResizeHandles } from '../../features/grid-editor/components/Grid/ResizeHandles';
+export { IsometricPreview } from '../../features/grid-editor/components/Grid/IsometricPreview';
+export { QuickLabelPopover } from '../../features/grid-editor/components/Grid/QuickLabelPopover';
