+++ src/features/grid-editor/components/Grid/Bin.tsx
@@ -1,18 +1,18 @@
 import type { PointerEvent } from 'react';
 import { memo, useRef, useState, useCallback, useMemo } from 'react';
 import { useShallow } from 'zustand/shallow';
-import type { Bin as BinType, Category, Layer, Drawer, ResizeHandle } from '../../core/types';
+import type { Bin as BinType, Category, Layer, Drawer, ResizeHandle } from '../../../../core/types';
 import {
   useLayoutStore,
   useSelectionStore,
   useViewStore,
   useInteractionStore,
   useMobileStore,
-} from '../../core/store';
-import { useToastStore } from '../../core/store/toast';
-import { useResponsive } from '../../hooks';
-import { calcMaxGridUnits, DEFAULT_CATEGORY_COLOR } from '../../core/constants';
-import { getBinTextColors } from '../../shared/utils';
+} from '../../../../core/store';
+import { useToastStore } from '../../../../core/store/toast';
+import { useResponsive } from '../../../../hooks';
+import { calcMaxGridUnits, DEFAULT_CATEGORY_COLOR } from '../../../../core/constants';
+import { getBinTextColors } from '../../../../shared/utils';
 import { ResizeHandles } from './ResizeHandles';
 
 /** Clamp a value between min and max */
