+++ src/features/grid-editor/components/Grid/IsometricPreview.tsx
@@ -1,15 +1,15 @@
 import { useMemo, useCallback, useRef, useState, useEffect } from "react"
 import { Canvas } from "@react-three/fiber"
 import { useShallow } from "zustand/shallow"
-import { useLayoutStore, useUIStore } from "../../core/store"
+import { useLayoutStore, useUIStore } from "../../../../core/store"
 import {
   STAGING_ID,
   DEFAULT_CATEGORY_COLOR,
   calcMaxGridUnits,
-} from "../../core/constants"
-import { useResponsive } from "../../shared/hooks"
-import { use3DPreviewKeyboard } from "../../hooks/use3DPreviewKeyboard"
-import { getLayerZStart } from "../../utils/collision"
+} from "../../../../core/constants"
+import { useResponsive } from "../../../../shared/hooks"
+import { use3DPreviewKeyboard } from "../../../../hooks/use3DPreviewKeyboard"
+import { getLayerZStart } from "../../../../utils/collision"
 import { Scene, type SceneHandle } from "./IsometricPreview/Scene"
 import { BinMesh } from "./IsometricPreview/BinMesh"
 import { SplitLineOverlay } from "./IsometricPreview/SplitLineOverlay"
