+++ src/features/grid-editor/components/Grid/IsometricPreview/BatchedCornerMarkers.tsx
