+++ src/features/grid-editor/components/Grid/IsometricPreview/BinCornerMarkers.tsx
