+++ src/features/grid-editor/components/Grid/IsometricPreview/BinMesh.tsx
@@ -1,8 +1,8 @@
 import { useRef } from 'react';
 import * as THREE from 'three';
 import { useFrame } from '@react-three/fiber';
-import type { Bin } from '../../../core/types';
-import { useBinGeometry } from '../../../hooks/useBinGeometry';
+import type { Bin } from '../../../../../core/types';
+import { useBinGeometry } from '../../../../../hooks/useBinGeometry';
 
 interface BinMeshProps {
   bin: Bin;
