+++ src/features/grid-editor/components/Grid/IsometricPreview/DrawerDimensions.tsx
