+++ src/features/grid-editor/components/Grid/IsometricPreview/FloorGrid.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import * as THREE from 'three';
-import type { FractionalEdge } from '../../../core/types';
+import type { FractionalEdge } from '../../../../../core/types';
 
 interface FloorGridProps {
   width: number;
