+++ src/features/grid-editor/components/Grid/IsometricPreview/FrontLabel.tsx
