+++ src/features/grid-editor/components/Grid/IsometricPreview/ScaleIndicator.tsx
