+++ src/features/grid-editor/components/Grid/IsometricPreview/SplitLineOverlay.tsx
