+++ src/features/grid-editor/components/Grid/ResizeHandle.tsx
@@ -1,7 +1,7 @@
 import type { PointerEvent } from 'react';
 import { memo } from 'react';
-import type { ResizeHandle as ResizeHandleType, HandleVariant, HandlePlacement } from '../../core/types';
-import { getHandlePosition, getHandleVisual, isCornerHandle } from '../../utils/handlePositioning';
+import type { ResizeHandle as ResizeHandleType, HandleVariant, HandlePlacement } from '../../../../core/types';
+import { getHandlePosition, getHandleVisual, isCornerHandle } from '../../../../utils/handlePositioning';
 
 interface ResizeHandleProps {
   handle: ResizeHandleType;
