+++ src/features/grid-editor/components/Grid/ResizeHandles.tsx
@@ -1,8 +1,8 @@
 import type { PointerEvent } from 'react';
 import { memo } from 'react';
-import type { ResizeHandle as ResizeHandleType, HandleVariant, HandlePlacement } from '../../core/types';
+import type { ResizeHandle as ResizeHandleType, HandleVariant, HandlePlacement } from '../../../../core/types';
 import { ResizeHandle } from './ResizeHandle';
-import { getAllHandles, shouldUseExternalHandles } from '../../utils/handlePositioning';
+import { getAllHandles, shouldUseExternalHandles } from '../../../../utils/handlePositioning';
 
 interface ResizeHandlesProps {
   binWidth: number;
